Get rid of async wasm recrypt interface
