+++ package.json
@@ -58,7 +58,7 @@
         "ts-jest": "^25.3.0",
         "typescript": "3.8.3",
         "typestrict": "^1.0.0",
-        "webpack": "^4.42.1",
+        "webpack": "^5.38.1",
         "webpack-cli": "^3.3.11",
         "webpack-dev-server": "^3.11.0",
         "worker-loader": "^2.0.0"
