+++ src/frame/worker/SearchCrypto.ts
@@ -1,28 +1,16 @@
-import Future from "futurejs";
-import {ErrorCodes} from "../../Constants";
-import SDKError from "../../lib/SDKError";
-import loadRecrypt from "./crypto/recrypt";
+import * as Recrypt from "./crypto/recrypt";
 
 /**
- * Tokenize the provided data with padding and map to the proper error code on failure.
+ * Tokenize the provided data with padding.
  */
-export const tokenizeData = (data: string, salt: Uint8Array, partitionId?: string): Future<SDKError, Uint32Array> =>
-    loadRecrypt()
-        .map((recrypt) => recrypt.tokenizeData(data, salt, partitionId))
-        .errorMap((e) => new SDKError(e, ErrorCodes.SEARCH_TOKENIZE_DATA_FAILURE));
+export const tokenizeData = (data: string, salt: Uint8Array, partitionId?: string): Uint32Array => Recrypt.tokenizeData(data, salt, partitionId);
 
 /**
  * Tokenize the provided query and map to the proper error code on failure.
  */
-export const tokenizeQuery = (query: string, salt: Uint8Array, partitionId?: string): Future<SDKError, Uint32Array> =>
-    loadRecrypt()
-        .map((recrypt) => recrypt.tokenizeQuery(query, salt, partitionId))
-        .errorMap((e) => new SDKError(e, ErrorCodes.SEARCH_TOKENIZE_QUERY_FAILURE));
+export const tokenizeQuery = (query: string, salt: Uint8Array, partitionId?: string): Uint32Array => Recrypt.tokenizeQuery(query, salt, partitionId);
 
 /**
  * Load Recrypt and then transliterate the provided string.
  */
-export const transliterateString = (data: string): Future<SDKError, string> =>
-    loadRecrypt()
-        .errorMap((e) => new SDKError(e, ErrorCodes.WEBASSEMBLY_SUPPORT_FAILURE))
-        .map((recrypt) => recrypt.transliterateString(data));
+export const transliterateString = (data: string): string => Recrypt.transliterateString(data);
