+++ src/frame/worker/tests/UserCrypto.test.ts
@@ -2,7 +2,7 @@ import Future from "futurejs";
 import {ErrorCodes} from "../../../Constants";
 import * as TestUtils from "../../../tests/TestUtils";
 import * as AES from "../crypto/aes";
-import * as Recrypt from "../crypto/recrypt/RecryptWasm";
+import * as Recrypt from "../crypto/recrypt";
 import * as UserCrypto from "../UserCrypto";
 
 describe("UserCrypto", () => {
@@ -309,36 +309,14 @@ describe("UserCrypto", () => {
                 authHeaderSignature: "sig2",
             });
 
-            UserCrypto.signRequestPayload(1, "user-10", TestUtils.getSigningKeyPair(), "GET", "/path/to/resource", "body").engage(
-                (e) => fail(e.message),
-                (result) => {
-                    expect(result).toEqual({
-                        userContextHeader: "comma,list,stuff",
-                        requestHeaderSignature: "sig1",
-                        authHeaderSignature: "sig2",
-                    });
-                    expect(Recrypt.createRequestSignature).toHaveBeenCalledWith(
-                        1,
-                        "user-10",
-                        TestUtils.getSigningKeyPair(),
-                        "GET",
-                        "/path/to/resource",
-                        "body"
-                    );
-                }
-            );
-        });
+            const result = UserCrypto.signRequestPayload(1, "user-10", TestUtils.getSigningKeyPair(), "GET", "/path/to/resource", "body");
 
-        it("maps a recrypt error into an SDKError", () => {
-            spyOn(Recrypt, "createRequestSignature").and.throwError("REEFER MADNESS");
-
-            UserCrypto.signRequestPayload(1, "user-10", TestUtils.getSigningKeyPair(), "GET", "/path/to/resource", "body").engage(
-                (e) => {
-                    expect(e.message).toBe("REEFER MADNESS");
-                    expect(e.code).toBe(ErrorCodes.SIGNATURE_GENERATION_FAILURE);
-                },
-                (_) => fail("shouldn't return if recrypt throws")
-            );
+            expect(result).toEqual({
+                userContextHeader: "comma,list,stuff",
+                requestHeaderSignature: "sig1",
+                authHeaderSignature: "sig2",
+            });
+            expect(Recrypt.createRequestSignature).toHaveBeenCalledWith(1, "user-10", TestUtils.getSigningKeyPair(), "GET", "/path/to/resource", "body");
         });
     });
 });
